requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
pinecone-client==2.2.4
pdfminer.six==20221105
python-docx==1.1.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
tiktoken==0.5.2
